Namespaced command actions into Robotics::Actions module.
Changed the `:action_exist?` validation to check if if action is defined as private method in Robotics::Actions - this is for ensuring that only defined actions would be run, and not other object methods.

Added a spec to check that, for example, a `:run` action in a command wouldn't trigger the `:run` method execution.

Fixed Rubocop's CaseIndentation namespace from `Style` to `Layout` to get rid of a warning.

Tested the app on Windows with Ruby 2.4.1